refactor: moved nested routes as folder index.tsx
