+++ fresh.gen.ts
@@ -6,25 +6,25 @@ import config from "./deno.json" assert { type: "json" };
 import * as $0 from "./routes/_404.tsx";
 import * as $1 from "./routes/_500.tsx";
 import * as $2 from "./routes/api/joke.ts";
-import * as $3 from "./routes/blog.tsx";
-import * as $4 from "./routes/blog/how-create-theme-switcher-deno-fresh.tsx";
+import * as $3 from "./routes/blog/how-create-theme-switcher-deno-fresh.tsx";
+import * as $4 from "./routes/blog/index.tsx";
 import * as $5 from "./routes/blog/stopping-theme-flickering-deno-fresh.tsx";
 import * as $6 from "./routes/certificates.tsx";
 import * as $7 from "./routes/index.tsx";
 import * as $8 from "./routes/me.tsx";
-import * as $9 from "./routes/projects.tsx";
-import * as $10 from "./routes/projects/expenses-tracker.tsx";
-import * as $11 from "./routes/projects/food-order.tsx";
-import * as $12 from "./routes/tools.tsx";
-import * as $13 from "./routes/tools/highlighted-text/[text].tsx";
+import * as $9 from "./routes/projects/expenses-tracker.tsx";
+import * as $10 from "./routes/projects/food-order.tsx";
+import * as $11 from "./routes/projects/index.tsx";
+import * as $12 from "./routes/tools/highlighted-text/[text].tsx";
+import * as $13 from "./routes/tools/index.tsx";
 import * as $14 from "./routes/tools/retirement-calculator.tsx";
 import * as $15 from "./routes/tools/syntax-highlight.tsx";
-import * as $16 from "./routes/toys.tsx";
+import * as $16 from "./routes/toys/index.tsx";
 import * as $17 from "./routes/toys/insanity.tsx";
 import * as $18 from "./routes/toys/spinners.tsx";
 import * as $19 from "./routes/what-is-this.tsx";
-import * as $20 from "./routes/work.tsx";
-import * as $21 from "./routes/work/form.tsx";
+import * as $20 from "./routes/work/form.tsx";
+import * as $21 from "./routes/work/index.tsx";
 import * as $$0 from "./islands/AddNewExpense.tsx";
 import * as $$1 from "./islands/Collapsible.tsx";
 import * as $$2 from "./islands/CopyTextAreaToClipboard.tsx";
@@ -44,25 +44,25 @@ const manifest = {
     "./routes/_404.tsx": $0,
     "./routes/_500.tsx": $1,
     "./routes/api/joke.ts": $2,
-    "./routes/blog.tsx": $3,
-    "./routes/blog/how-create-theme-switcher-deno-fresh.tsx": $4,
+    "./routes/blog/how-create-theme-switcher-deno-fresh.tsx": $3,
+    "./routes/blog/index.tsx": $4,
     "./routes/blog/stopping-theme-flickering-deno-fresh.tsx": $5,
     "./routes/certificates.tsx": $6,
     "./routes/index.tsx": $7,
     "./routes/me.tsx": $8,
-    "./routes/projects.tsx": $9,
-    "./routes/projects/expenses-tracker.tsx": $10,
-    "./routes/projects/food-order.tsx": $11,
-    "./routes/tools.tsx": $12,
-    "./routes/tools/highlighted-text/[text].tsx": $13,
+    "./routes/projects/expenses-tracker.tsx": $9,
+    "./routes/projects/food-order.tsx": $10,
+    "./routes/projects/index.tsx": $11,
+    "./routes/tools/highlighted-text/[text].tsx": $12,
+    "./routes/tools/index.tsx": $13,
     "./routes/tools/retirement-calculator.tsx": $14,
     "./routes/tools/syntax-highlight.tsx": $15,
-    "./routes/toys.tsx": $16,
+    "./routes/toys/index.tsx": $16,
     "./routes/toys/insanity.tsx": $17,
     "./routes/toys/spinners.tsx": $18,
     "./routes/what-is-this.tsx": $19,
-    "./routes/work.tsx": $20,
-    "./routes/work/form.tsx": $21,
+    "./routes/work/form.tsx": $20,
+    "./routes/work/index.tsx": $21,
   },
   islands: {
     "./islands/AddNewExpense.tsx": $$0,
