+++ routes/projects/index.tsx
@@ -1,14 +1,14 @@
 //? Head component with all Meta tags pre-set
-import { CustomHead } from "../components/base/CustomHead.tsx";
+import { CustomHead } from "../../components/base/CustomHead.tsx";
 //? Create blog content inside Base component
-import { Base } from "../components/base/Base.tsx";
+import { Base } from "../../components/base/Base.tsx";
 //? Default styled header
-import { StyledHeader } from "../components/UI/StyledHeader.tsx";
-import { GradientLink } from "../components/UI/GradientLink.tsx";
+import { StyledHeader } from "../../components/UI/StyledHeader.tsx";
+import { GradientLink } from "../../components/UI/GradientLink.tsx";
 //? Navigation Buttons to go back to the previous page or to the next article
-import { BlogNavigationButtons } from "../components/blog/BlogNavigationButtons.tsx";
+import { BlogNavigationButtons } from "../../components/blog/BlogNavigationButtons.tsx";
 //? Component to create a column grid of projects created
-import { ProjectsGrid } from "../components/projects/ProjectsGrid.tsx";
+import { ProjectsGrid } from "../../components/projects/ProjectsGrid.tsx";
 
 export default function Home() {
   return (
