+++ routes/tools/index.tsx
@@ -1,12 +1,12 @@
 //? Import CustomHead with appropriate metadata
-import { CustomHead } from "../components/base/CustomHead.tsx";
+import { CustomHead } from "../../components/base/CustomHead.tsx";
 //? Create blog content inside Base component
-import { Base } from "../components/base/Base.tsx";
+import { Base } from "../../components/base/Base.tsx";
 //? Default styled header
-import { StyledHeader } from "../components/UI/StyledHeader.tsx";
+import { StyledHeader } from "../../components/UI/StyledHeader.tsx";
 //? Navigation Buttons to go back to the previous page or to the next page (optional)
-import { BlogNavigationButtons } from "../components/blog/BlogNavigationButtons.tsx";
-import { GradientLink } from "../components/UI/GradientLink.tsx";
+import { BlogNavigationButtons } from "../../components/blog/BlogNavigationButtons.tsx";
+import { GradientLink } from "../../components/UI/GradientLink.tsx";
 
 export default function Home() {
   return (
